Update for geometry order
