+++ wrappers/Python/mui4py/geometry.cpp
@@ -11,6 +11,9 @@ void declare_geometry_shape(py::module &m)
 {
     std::string name = "_geometry_shape_" + config_name<Tconfig>();
     py::class_<mui::geometry::shape<Tconfig>>(m, name.c_str());
+
+    py::implicitly_convertible<mui::geometry::shape<Tconfig>,
+                               mui::geometry::any_shape<Tconfig>>();
 }
 
 template <typename Tconfig>
@@ -52,6 +55,20 @@ void declare_geometry_box(py::module &m)
 void geometry(py::module &m)
 {
 #ifdef PYTHON_INT_64
+    declare_geometry_any_shape<mui::mui_config_1dx>(m);
+    declare_geometry_any_shape<mui::mui_config_2dx>(m);
+    declare_geometry_any_shape<mui::mui_config_3dx>(m);
+    declare_geometry_any_shape<mui::mui_config_1fx>(m);
+    declare_geometry_any_shape<mui::mui_config_2fx>(m);
+    declare_geometry_any_shape<mui::mui_config_3fx>(m);
+
+    declare_geometry_shape<mui::mui_config_1dx>(m);
+    declare_geometry_shape<mui::mui_config_2dx>(m);
+    declare_geometry_shape<mui::mui_config_3dx>(m);
+    declare_geometry_shape<mui::mui_config_1fx>(m);
+    declare_geometry_shape<mui::mui_config_2fx>(m);
+    declare_geometry_shape<mui::mui_config_3fx>(m);
+
     declare_geometry_point<mui::mui_config_1dx>(m);
     declare_geometry_point<mui::mui_config_2dx>(m);
     declare_geometry_point<mui::mui_config_3dx>(m);
@@ -73,21 +90,21 @@ void geometry(py::module &m)
     declare_geometry_sphere<mui::mui_config_2fx>(m);
     declare_geometry_sphere<mui::mui_config_3fx>(m);
 
-    declare_geometry_any_shape<mui::mui_config_1dx>(m);
-    declare_geometry_any_shape<mui::mui_config_2dx>(m);
-    declare_geometry_any_shape<mui::mui_config_3dx>(m);
-    declare_geometry_any_shape<mui::mui_config_1fx>(m);
-    declare_geometry_any_shape<mui::mui_config_2fx>(m);
-    declare_geometry_any_shape<mui::mui_config_3fx>(m);
-
-    declare_geometry_shape<mui::mui_config_1dx>(m);
-    declare_geometry_shape<mui::mui_config_2dx>(m);
-    declare_geometry_shape<mui::mui_config_3dx>(m);
-    declare_geometry_shape<mui::mui_config_1fx>(m);
-    declare_geometry_shape<mui::mui_config_2fx>(m);
-    declare_geometry_shape<mui::mui_config_3fx>(m);
-
 #elif defined PYTHON_INT_32
+
+    declare_geometry_any_shape<mui::mui_config_1d>(m);
+    declare_geometry_any_shape<mui::mui_config_2d>(m);
+    declare_geometry_any_shape<mui::mui_config_3d>(m);
+    declare_geometry_any_shape<mui::mui_config_1f>(m);
+    declare_geometry_any_shape<mui::mui_config_2f>(m);
+    declare_geometry_any_shape<mui::mui_config_3f>(m);
+
+    declare_geometry_shape<mui::mui_config_1d>(m);
+    declare_geometry_shape<mui::mui_config_2d>(m);
+    declare_geometry_shape<mui::mui_config_3d>(m);
+    declare_geometry_shape<mui::mui_config_1f>(m);
+    declare_geometry_shape<mui::mui_config_2f>(m);
+    declare_geometry_shape<mui::mui_config_3f>(m);
 
     declare_geometry_point<mui::mui_config_1d>(m);
     declare_geometry_point<mui::mui_config_2d>(m);
@@ -109,20 +126,6 @@ void geometry(py::module &m)
     declare_geometry_sphere<mui::mui_config_1f>(m);
     declare_geometry_sphere<mui::mui_config_2f>(m);
     declare_geometry_sphere<mui::mui_config_3f>(m);
-
-    declare_geometry_any_shape<mui::mui_config_1d>(m);
-    declare_geometry_any_shape<mui::mui_config_2d>(m);
-    declare_geometry_any_shape<mui::mui_config_3d>(m);
-    declare_geometry_any_shape<mui::mui_config_1f>(m);
-    declare_geometry_any_shape<mui::mui_config_2f>(m);
-    declare_geometry_any_shape<mui::mui_config_3f>(m);
-
-    declare_geometry_shape<mui::mui_config_1d>(m);
-    declare_geometry_shape<mui::mui_config_2d>(m);
-    declare_geometry_shape<mui::mui_config_3d>(m);
-    declare_geometry_shape<mui::mui_config_1f>(m);
-    declare_geometry_shape<mui::mui_config_2f>(m);
-    declare_geometry_shape<mui::mui_config_3f>(m);
 
 #else
 #error PYTHON_INT_[32|64] not defined.
